+++ CHANGELOG.md
@@ -24,6 +24,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to use names with trailing `_` without a reason
 - Forbids to use `super()` with arguments or outside of methods
 - Forbids to have too many `except` cases
+- Enforces to have an empty line after `coding` comment
 
 ### Bugfixes